requirements.txt
groq
python-dotenv
fastapi
uvicorn[standard]
httpx
google-genai
qdrant-client[fastembed]
pymupdf
tqdm
kaggle
